pyproject.toml
[project]
name = "rossmann-store"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "mlflow>=3.3.2",
    "notebook>=7.4.5",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]


[tool.uv.sources]
utils = { path = "src/utils" }
